requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-cors-headers==4.1.0
django-filter==23.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
importlib-metadata==6.6.0
Markdown==3.4.3
pycparser==2.21
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
zipp==3.15.0
pillow==9.5.0
